refactor: consolidate edge function helpers
