--- supabase/functions/_shared/crypto.ts
@@ -1,13 +1,0 @@
-/** Shared WebCrypto helpers for edge functions. */
-
-/** Cryptographically random URL-safe base64 string from `byteCount` bytes. */
-export function randomBase64Url(byteCount: number): string {
-  const bytes = new Uint8Array(byteCount);
-  crypto.getRandomValues(bytes);
-  let binary = '';
-  for (const byte of bytes) binary += String.fromCharCode(byte);
-  return btoa(binary)
-    .replaceAll('+', '-')
-    .replaceAll('/', '_')
-    .replace(/=+$/, '');
-}
